# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation untuk API Manajemen Senjata",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/jenissenjata": {
            "get": {
                "tags": [
                    "JenisSenjata"
                ],
                "summary": "Menampilkan seluruh data jenis senjata",
                "operationId": "f02672db7fc2acb405b752b60fd314fc",
                "responses": {
                    "200": {
                        "description": "Berhasil menampilkan semua jenis senjata"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "JenisSenjata"
                ],
                "summary": "Membuat jenis senjata baru",
                "operationId": "f13ff059d50caf60896533a4690886c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_jenis"
                                ],
                                "properties": {
                                    "nama_jenis": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jenis senjata berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/jenissenjata/{id}": {
            "get": {
                "tags": [
                    "JenisSenjata"
                ],
                "summary": "Menampilkan detail jenis senjata berdasarkan ID",
                "operationId": "2051df95d0e5035802e483a6616d3c4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data jenis senjata ditemukan"
                    },
                    "404": {
                        "description": "Jenis senjata tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "JenisSenjata"
                ],
                "summary": "Mengubah data jenis senjata berdasarkan ID",
                "operationId": "45eb9154c4590429412dbf44ca4f112c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_jenis"
                                ],
                                "properties": {
                                    "nama_jenis": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jenis senjata berhasil diperbarui"
                    },
                    "404": {
                        "description": "Jenis senjata tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "JenisSenjata"
                ],
                "summary": "Menghapus jenis senjata berdasarkan ID",
                "operationId": "77978d82d74da2a485c00dde20459737",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jenis senjata berhasil dihapus"
                    },
                    "404": {
                        "description": "Jenis senjata tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "1": {
            "sanctum": []
        }
    }
}